refactor: 생성자에서 price 타입 int에서 long으로 변경

+++ src/main/java/domain/LottoRank.java
@@ -14,7 +14,7 @@ public enum LottoRank {
     private final long price;
     private final boolean matchBonus;
 
-    LottoRank(int matchCount, int price, boolean matchBonus) {
+    LottoRank(int matchCount, long price, boolean matchBonus) {
         this.matchCount = matchCount;
         this.price = price;
         this.matchBonus = matchBonus;
